第十一次提交 day10完成
ElasticSearch检索
聚合
aggs
  ""
    terms

嵌入式
  ""
    nested
    aggs

高亮
  HighlightBuilder()
  .field("").preTags("").postTags("")

  HighlightQuery()
  setHighlightQuery()

聚合
  NativeSearchQuery

  addAggregation()  添加聚合

  subAggregation()  子聚合

  AggregationBuilders.terms("聚合名").field("聚合字段").size(长度)

嵌入式聚合
  AggregationBuilders.nested("聚合名","嵌入式名")

  将聚合作为嵌入式子聚合

聚合结果处理
  getAggregations().get("聚合名")

  ParsedNested ParsedLongTerms... 转为具体类型 方法多些

  getBuckets() 获取所有桶

  getKeyAsString() 获取键 String类型
  getKeyAsNumber() 获取键  数值类型

+++ model/src/main/java/com/atguigu/gmall/model/vo/GoodsSearchResultVo.java
@@ -16,6 +16,6 @@ public class GoodsSearchResultVo {
     private List<SearchResponseAttrVo> attrsList; //检索区 : 平台属性列表信息
     private SearchResponseOmVo orderMap; //排序条件
     private List<Goods> goodsList; //检索到的所有商品集合
-    private Long pageNo;  //当前页
-    private Long totalPages;  // 总页码
+    private Integer pageNo;  //当前页
+    private Integer totalPages;  // 总页码
 }
